The namespace is divided between different files, for each class. With these changes, the TaskPersister helper methods are encapsulated.
